fix bug long value of file length

+++ android/src/main/java/com/stringee/stringeeflutterplugin/StringeeFlutterPlugin.java
@@ -424,7 +424,8 @@ public class StringeeFlutterPlugin implements MethodCallHandler, EventChannel.St
                             if (msgMap.containsKey("filename"))
                                 message.setFileName((String) msgMap.get("filename"));
                             if (msgMap.containsKey("length"))
-                                message.setFileLength((Long) msgMap.get("length"));
+//                                message.setFileLength((Long) msgMap.get("length"));
+                            message.setFileLength(((Integer) msgMap.get("length")).longValue());
                             break;
                         case LOCATION:
                             message.setLatitude((Double) msgMap.get("lat"));
